+++ frb_example/pure_dart/rust/src/api/method.rs
@@ -131,32 +131,3 @@ impl SimpleStructTwinNormal {
         arg.into_iter().map(|x| x.one).collect()
     }
 }
-
-pub struct ConstructorTranslatableStructTwinNormal {
-    pub one: String,
-}
-
-impl ConstructorTranslatableStructTwinNormal {
-    pub fn new() -> Self {
-        Self {
-            one: "hello".to_owned(),
-        }
-    }
-}
-
-#[frb(opaque)]
-pub struct ConstructorOpaqueStructTwinNormal {
-    pub one: String,
-}
-
-impl ConstructorOpaqueStructTwinNormal {
-    pub fn new() -> Self {
-        Self {
-            one: "hello".to_owned(),
-        }
-    }
-
-    pub fn check(&self) {
-        assert_eq!(self.one, "hello");
-    }
-}
