+++ frb_example/pure_dart/rust/src/api/mod.rs
@@ -6,6 +6,7 @@ pub mod attribute;
 pub mod benchmark_misc;
 pub mod chrono_type;
 pub mod comment;
+pub mod constructor;
 pub mod customization;
 pub mod dart_dynamic;
 pub mod dart_fn;
